pyproject.toml
[tool.poetry]
name = "silverrail"
version = "0.1.0"
description = ""
authors = ["Ash Omaraie <49348213+aomarai@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"

[tool.poetry.dev-dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
